run_erp_suite.sh: Remove timestamps from apt cache
Calculate a unique build id based on the build number but also the
contents of the local apt-cache, so that when the available set of
packages changes, the build id changes. Remove "^ Time" from the results
because the dpkg status file seems to be dynamically generated and
contains a timestamp.

Also, reorder test plans to have performance be last.

Signed-off-by: Dan Rue <dan.rue@linaro.org>